add form validation on client
