+++ src/app/api/uploadthing/core.ts
@@ -20,7 +20,6 @@ export const ourFileRouter = {
       // If you throw, the user will not be able to upload
       // eslint-disable-next-line @typescript-eslint/only-throw-error
       if (!user) throw new UploadThingError("Unauthorized");
-      console.log(input, "test test");
       // Whatever is returned here is accessible in onUploadComplete as `metadata`
       return { userId: user.userId, input };
     })
