+++ src/components/investment_form.tsx
@@ -1,0 +1,333 @@
+"use client";
+import React, { useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
+import {
+  Dialog,
+  DialogContentNoClose as DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "~/components/ui/dialog";
+import { Button } from "~/components/ui/button";
+import Link from "next/link";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "~/components/ui/form";
+import { Input } from "~/components/ui/input";
+import { Checkbox } from "~/components/ui/checkbox";
+import { createInvestment } from "~/server/action/create_investment";
+import { z } from "zod";
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { type business } from "~/server/db/schema";
+import { isCreditCard, isDate } from "validator";
+
+function DialogCountdown({ isFormValid }: { isFormValid: boolean }) {
+  const [countdown, setCountdown] = useState(3);
+  const [startCountdown, setStartCountdown] = useState(false);
+
+  const handleConfirmInvestment = () => {
+    setStartCountdown(true);
+  };
+
+  const router = useRouter();
+  useEffect(() => {
+    let timer: NodeJS.Timeout;
+    if (startCountdown && countdown > 0) {
+      timer = setTimeout(() => setCountdown(countdown - 1), 1000);
+      return () => clearTimeout(timer);
+    } else if (countdown === 0) {
+      router.push("/investor_portfolio");
+    }
+    return () => clearTimeout(timer);
+  }, [countdown, router, startCountdown]);
+
+  return (
+    <Dialog>
+      <DialogTrigger asChild>
+        <Button
+          type="submit"
+          className="mt-6 w-full bg-blue-500 py-3 text-lg font-semibold text-white hover:bg-blue-600 disabled:cursor-not-allowed disabled:opacity-50"
+          onClick={handleConfirmInvestment}
+          disabled={!isFormValid}
+        >
+          Confirm investment
+        </Button>
+      </DialogTrigger>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>Investing request has been sent.</DialogTitle>
+          <DialogDescription>
+            Redirecting you to your portfolio in {countdown} seconds.
+            <Link href="/investor_portfolio">
+              <Button className="mt-6 w-full bg-blue-500 py-3 text-lg font-semibold text-white hover:bg-blue-600">
+                Go to your portfolio
+              </Button>
+            </Link>
+          </DialogDescription>
+        </DialogHeader>
+      </DialogContent>
+    </Dialog>
+  );
+}
+
+export function InvestingForm({
+  businessData,
+}: {
+  businessData: typeof business.$inferSelect;
+}) {
+  const createInvestmentBind = createInvestment.bind(
+    null,
+    businessData.businessID,
+  );
+
+  const FormSchema = z.object({
+    terms: z.boolean().default(false),
+
+    cardNumber: z
+      .string()
+      .min(1, { message: "Credit card number is required." })
+      .refine((value) => isCreditCard(value.replace(/\s/g, "")), {
+        message: "Please enter a valid credit card number.",
+      })
+      .refine(
+        (value) => {
+          const digits = value.replace(/\s/g, "");
+          return digits.length >= 13 && digits.length <= 16;
+        },
+        {
+          message: "Credit card number must be between 13 and 16 digits.",
+        },
+      ),
+
+    expirationDate: z
+      .string()
+      .min(1, { message: "Expiration date is required." })
+      .refine(
+        (value) => {
+          const [month, year] = value.split("/");
+          if (!month || !year || month.length !== 2 || year.length !== 2)
+            return false;
+
+          const currentDate = new Date();
+          const expDate = new Date(2000 + parseInt(year), parseInt(month) - 1);
+
+          return !isNaN(expDate.getTime()) && expDate > currentDate;
+        },
+        {
+          message: "Please enter a valid expiration date (MM/YY)",
+        },
+      ),
+
+    cvv: z
+      .string()
+      .min(1, { message: "CVV is required." })
+      .refine((value) => /^\d{3,4}$/.test(value), {
+        message: "Please enter a valid CVV (3-4 digits)",
+      }),
+
+    amount: z
+      .number({
+        required_error: "Amount is required.",
+        invalid_type_error: "Amount must be a number.",
+      })
+      .min(businessData.min_investment!, {
+        message: `Minimum investment amount is ${businessData.min_investment} $`,
+      })
+      .positive({
+        message: "Amount must be positive.",
+      }),
+  });
+
+  const form = useForm<z.infer<typeof FormSchema>>({
+    resolver: zodResolver(FormSchema),
+    defaultValues: {
+      terms: false,
+      cardNumber: "",
+      expirationDate: "",
+      cvv: "",
+    },
+    mode: "onBlur",
+  });
+
+  const isFormValid = form.formState.isValid && form.watch("terms");
+
+  return (
+    <div className="font-geist-sans mx-auto max-w-lg pb-20">
+      <div className="my-6 flex flex-row items-center justify-center gap-2">
+        <img
+          src="https://utfs.io/f/bb1dabab-7c7c-40d7-8ea5-030fdc7f1d96-ny8zu1.jpg"
+          alt="B2D Ventures Logo"
+          className="h-[60px] w-[60px]"
+        />
+        <h1 className="text-4xl font-bold">Invest in {businessData.company}</h1>
+      </div>
+      <div className="mt-6">
+        <h2 className="text-lg font-bold">Investment amount</h2>
+        <p className="mb-4 text-sm text-gray-600">
+          Payments are processed immediately.
+        </p>
+        <Form {...form}>
+          <form action={createInvestmentBind} className="space-y-6">
+            <FormField
+              control={form.control}
+              name="amount"
+              render={({ field }) => (
+                <FormItem>
+                  <FormControl>
+                    <Input
+                      placeholder={`Minimum investment ${businessData.min_investment!.toLocaleString()} $`}
+                      className="p-4 text-lg"
+                      type="number"
+                      {...field}
+                      onChange={(e) => field.onChange(Number(e.target.value))}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+
+            <div className="mt-8">
+              <h2 className="text-lg font-bold">Payment information</h2>
+
+              <div className="mt-4 space-y-4">
+                <FormField
+                  control={form.control}
+                  name="cardNumber"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormControl>
+                        <Input
+                          placeholder="Card number"
+                          className="p-3 text-lg"
+                          {...field}
+                        />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+
+                <div className="flex gap-2">
+                  <FormField
+                    control={form.control}
+                    name="expirationDate"
+                    render={({ field }) => (
+                      <FormItem className="flex-1">
+                        <FormControl>
+                          <Input
+                            placeholder="MM/YY"
+                            className="p-3 text-lg"
+                            {...field}
+                            onChange={(e) => {
+                              let value = e.target.value;
+                              if (value.length === 2 && !value.includes("/")) {
+                                value += "/";
+                              }
+                              field.onChange(value);
+                            }}
+                            maxLength={5}
+                          />
+                        </FormControl>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="cvv"
+                    render={({ field }) => (
+                      <FormItem className="w-1/3">
+                        <FormControl>
+                          <Input
+                            placeholder="CVV"
+                            className="p-3 text-lg"
+                            type="password"
+                            maxLength={4}
+                            {...field}
+                          />
+                        </FormControl>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+                </div>
+              </div>
+            </div>
+            <div className="mt-6">
+              <h2 className="text-lg font-semibold">Terms</h2>
+              <div className="mt-4 space-y-2 rounded-md bg-gray-100 p-4 text-sm text-gray-700">
+                <p>
+                  I understand that I can cancel my investment up until 10/30/24
+                  (48 hours prior to the deal deadline).
+                </p>
+                <p>
+                  I understand that Republic will receive a cash and securities
+                  commission as further detailed in the offering documents.
+                </p>
+                <p>
+                  I understand that investing this amount into several deals
+                  would better diversify my risk.
+                </p>
+                <p>
+                  I consent to electronic delivery of all documents, notices and
+                  agreements as related to my investment.
+                </p>
+                <p>
+                  I understand my investment won&apost be transferable for 12
+                  months and may not have a market for resale.
+                </p>
+                <p>
+                  I have read the educational materials and agree to the Terms
+                  of Service, including arbitration provisions.
+                </p>
+                <p>
+                  I understand this investment is risky and that I shouldn&apost
+                  invest unless I can afford to lose all invested funds.
+                </p>
+                <p>
+                  I understand I am responsible for all fees and charges
+                  associated with the use of my payment method.
+                </p>
+                <p>
+                  I confirm that this investment, together with all my other
+                  Regulation Crowdfunding investments during the past 12 months
+                  on any crowdfunding portal, does not exceed my investment
+                  limit.
+                </p>
+              </div>
+            </div>
+
+            <FormField
+              control={form.control}
+              name="terms"
+              render={({ field }) => (
+                <FormItem className="mt-4 flex items-center space-x-3">
+                  <FormControl>
+                    <Checkbox
+                      checked={field.value}
+                      onCheckedChange={field.onChange}
+                      id="terms"
+                    />
+                  </FormControl>
+                  <FormLabel>
+                    I have read and accept the terms of investment
+                  </FormLabel>
+                </FormItem>
+              )}
+            />
+            <DialogCountdown isFormValid={isFormValid} />
+          </form>
+        </Form>
+      </div>
+    </div>
+  );
+}
